v0.9 Leaderboard places and test fill in the choice

+++ app/src/main/java/com/example/duelingo/dto/response/UserInLeaderboardResponse.kt
@@ -5,5 +5,5 @@ data class UserInLeaderboardResponse(
     val username: String,
     val points: Int,
     val avatarUrl: String,
-    val rank: Int
+    val rank: Long
 )
